Resolve precomp symbols using dlsym() if dynamically linked
When :runtime is included as a direct dependency of a test, Bazel
chooses to optimize the linking phase by dynamically linking with
several .so files instead of statically linking everything together. As
a result, the weak symbols are not overriden and we need to use dlsym()
to properly resolve them.